requirements-cpu.txt
absl-py==2.1.0
addict==2.4.0
albumentations==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
async-timeout==5.0.1
basicsr==1.4.2
beautifulsoup4==4.12.3
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
contourpy==1.3.1
customtkinter==5.2.0
cycler==0.12.1
Cython==3.0.11
darkdetect==0.8.0
dnspython==2.7.0
easydict==1.13
email_validator==2.2.0
exceptiongroup==1.2.2
facexlib==0.3.0
fastapi==0.115.7
fastapi-cli==0.0.7
filelock==3.16.1
filterpy==1.4.5
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.12.0
future==1.0.0
gast==0.4.0
gdown==5.2.0
gfpgan==1.3.8
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.69.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.36.1
insightface==0.7.3
Jinja2==3.1.5
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.8
kombu==5.4.2
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lmdb==1.6.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.14.0
onnxruntime==1.15.0
opencv-python==4.8.0.74
opencv-python-headless==4.10.0.84
opennsfw2==0.10.2
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
prettytable==3.12.0
prompt_toolkit==3.0.50
protobuf==4.23.4
psutil==5.9.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
qudida==0.0.4
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.15.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
sympy==1.13.1
tb-nightly==2.19.0a20250110
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tk==0.1.0
tkinterdnd2==0.3.0
tomli==2.2.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.65.0
triton==2.2.0
typer==0.15.1
typing_extensions>=4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.0
yapf==0.43.0
